Fix checkWinner
Problem was the board imported into screenController was the original
board and adding tokens was being done by screenController not
gameController. The result is gameController was processing a different
board than screenController.

Now, all the game logic happens in gameController. ScreenController Now
imports the board from gameController - they now deal with the same
board which alleviated the issue.