Add newGame and ability to resetBoard
resetBoard() clears the board of markers and resets it back to blank.
Add games variable to track number of games played.
newGame() calls resetBoard(), resets the turns played, and changes the
starting player depending if games played is even or odd.
Add event listener to Restart Game to call newGame() on click.